+++ scarletio/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '1.0.67'
+__version__ = '1.0.68'
 
 
 from .core import *